Corrected basename() bug #5982:
        return pRet;
Corrected the implementation of basename() (in 0.37-pre9) so that the check for 4 is done before the parsing (the check is to validate the string is a valid psp ms filename (they start with 'ms0:' which is 4 bytes long)).
Also made the return pointer static. But the fact that the pointer is returned by the function should not be a problem in this case, as it returns a substring of the passed string, so it doesn't really return local data.